updating process deposit function to be more readable and fork specific, breaking out components of pr 13944
